refactor: keep 0.56.2 modules below review threshold
